Revert "Moved util to weapon system folder."
This reverts commit 10b3066bfe4ad05ce84d4ad48fa64bc3d9b494a3.
